common: Notify condition variables while holding lock
It's a common recommendation to notify a condition variable after
releasing the associated lock. The goal being to optimize out pre-mature
wake up of a waiting thread only to find that the lock is unavailable
and then goes back to sleep.

Notifying a condition variable while holding the lock is perfectly valid
and doing otherwise is sometimes incorrect, so do the former.

I noticed this issue in a separate change that benchmarks `ThreadPool`.
While running on macOS, the threads were scheduled in a particular way
and that was accompanied by a spurious wake-up of a condition variable
that ultimately got used after being destroyed. This ultimately caused
the benchmark to hang:
https://github.com/whalbawi/spindle/runs/6083517564